Add user, regi_token and unauth_erorr serialiers, refactor quill_blog_posts_controller, refactor registration_tokens_controller

+++ app/lib/registration_token_life_cyclable.rb
@@ -20,6 +20,11 @@ module RegistrationTokenLifeCyclable
     def remove(token)
         registration_token = self.find_by_token(token)
 
-        registration_token.destroy
+        if registration_token
+            registration_token.destroy
+            true
+        else
+            false
+        end
     end
 end
